+++ src/pages/CreateDesign/components/Footer/index.tsx
@@ -1,8 +1,8 @@
 import { Button as MaterialButton } from '@material-ui/core';
 import { Button } from 'dumbs';
+import { currentStepAtom } from 'pages/CreateDesign/atom';
 import { FooterContainer } from 'pages/CreateDesign/components/Footer/Footer.css';
 import { useStepController } from 'pages/CreateDesign/components/Footer/hooks/useStepController';
-import { currentStepAtom } from 'pages/CreateDesign/recoils';
 import { PAGE } from 'pages/CreateDesign/types';
 import React from 'react';
 import { useRecoilValue } from 'recoil';
