+++ src/pages/CreateDesign/index.tsx
@@ -3,9 +3,9 @@ import Cover from 'pages/CreateDesign/Cover';
 import Outline from 'pages/CreateDesign/Outline';
 import Pattern from 'pages/CreateDesign/Pattern';
 import Review from 'pages/CreateDesign/Review';
+import { currentStepAtom } from 'pages/CreateDesign/atom';
 import Footer from 'pages/CreateDesign/components/Footer';
 import Header from 'pages/CreateDesign/components/Header';
-import { currentStepAtom } from 'pages/CreateDesign/recoils';
 import { PAGE } from 'pages/CreateDesign/types';
 import React from 'react';
 import { useRecoilValue } from 'recoil';
