+++ src/pages/MyInformation/components/Designs/index.tsx
@@ -3,8 +3,8 @@ import { FAILED_TO_GET_MY_DESIGNS } from 'constants/errors';
 import { List, Typography } from '@material-ui/core';
 import { useCommonSnackbar } from 'components/CommonSnackbar/useCommonSnackbar';
 import EmptyContent from 'dumbs/EmptyContent';
+import { tabItemLengthAtom } from 'pages/MyInformation/atom';
 import { useGetMyDesigns } from 'pages/MyInformation/hooks/useGetMyDesigns';
-import { tabItemLengthAtom } from 'pages/MyInformation/recoils';
 import { useEffect } from 'react';
 import InfiniteScroll from 'react-infinite-scroll-component';
 import { useSetRecoilState } from 'recoil';
