+++ src/pages/MyInformation/components/InformationTabs/useRenderEmptyContent.ts
@@ -1,5 +1,5 @@
 import { EmptyContentProps } from 'dumbs/EmptyContent';
-import { selectedTabAtom } from 'pages/MyInformation/recoils';
+import { selectedTabAtom } from 'pages/MyInformation/atom';
 import { DESIGN_MENU } from 'pages/MyInformation/types';
 import { useHistory } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
